styles: favicon, product hover added redirects
