+++ src/pages/Home/Home.jsx
@@ -1,5 +1,9 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 export default function HomePage() {
+  useEffect(() => {
+    document.title = "ecom | home";
+  }, []);
+
   return <>This is Home page</>;
 }
