+++ src/pages/ProductListing/productListing.css
@@ -77,7 +77,7 @@
 }
 
 .card:hover {
-  border: 4px solid white;
+  border: 0.2px solid gray;
   /* box-shadow: 5px 5px 5px 0px grey; */
 }
